move test TShardReader to a common place
