--- ydb/core/tx/columnshard/common/tests/shard_reader.cpp
@@ -1,4 +1,0 @@
-#include "shard_reader.h"
-
-namespace NKikimr::NOlap::NTests {
-}
